+++ apps/sb-mat-thm-demo/src/app/config/app-theme.config.ts
@@ -35,7 +35,7 @@ const _themeOptions: ThemeOption[] = [
     value: 'pastel',
     primaryColor: '#FFB7C5', // Light pink
     secondaryColor: '#FFDA61', // Deeper pastel yellow
-    tertiaryColor: '#7371FC', 
+    tertiaryColor: '#7371FC',
   }),
   ThemeOption.create({
     darkMode: 'dark',
@@ -43,7 +43,7 @@ const _themeOptions: ThemeOption[] = [
     value: 'deep-ocean',
     primaryColor: '#1E3A8A', // Deep blue
     secondaryColor: '#06B6D4', // Cyan
-    tertiaryColor: '#4E6E58', 
+    tertiaryColor: '#4E6E58',
   }),
   ThemeOption.create({
     darkMode: 'light',
@@ -51,7 +51,7 @@ const _themeOptions: ThemeOption[] = [
     value: 'cherry-blossom',
     primaryColor: '#EC4899', // Pink
     secondaryColor: '#84CC16', // Lime green
-    tertiaryColor: '#9BF3F0', 
+    tertiaryColor: '#9BF3F0',
   })
 ];
 
@@ -78,6 +78,7 @@ export const THEME_CONFIG = ThemingConfig.create({
   themeOptions: _themeOptions,
   defaultDarkModeType: 'dark',
   themeClassPrefix: 'my-app-theme',
+  darkModeClass: 'my-dark-mode',
   transitionOptions: {
     style: 'morph',
     // style: 'overlay',