+++ apps/sb-mat-thm-demo/src/app/sections/api/features/theme-service-api/theme-service-api.component_example.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { MatSelectModule } from '@angular/material/select';
 import { ThemeOption } from '@spider-baby/material-theming/config';
-import { ThemeService } from '@spider-baby/material-theming/service';
+import { SbThemeService } from '@spider-baby/material-theming/service';
 
 @Component({
   selector: 'sb-theme-demo-tester',