+++ libs/packages/@spider-baby/theming/components/src/lib/dark/dark-mode-toggle.component.ts
@@ -1,6 +1,5 @@
 import { NgTemplateOutlet } from '@angular/common';
 import { ChangeDetectionStrategy, Component, inject, input, output } from '@angular/core';
-import { ThemePalette } from '@angular/material/core';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { SbThemePalette } from '@spider-baby/material-theming/config';
@@ -15,7 +14,7 @@ import { SbThemeService } from '@spider-baby/material-theming/service';
  * 
  * @example
  * ```html
- * <sb-dark-mode-toggle-mat [showTooltip]="true"></sb-dark-mode-toggle-mat>
+ * <sb-dark-mode-toggle-mat [showTooltip]="true"/>
  * ```
  */
 @Component({
@@ -24,13 +23,13 @@ import { SbThemeService } from '@spider-baby/material-theming/service';
   imports: [
     NgTemplateOutlet,
     MatTooltipModule,
-    MatSlideToggleModule
+    MatSlideToggleModule,
   ],
   templateUrl: './dark-mode-toggle.component.html',
   styleUrls: ['./dark-mode-toggle.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
   host: {
-    '[class]': '_color()',
+    '[style.color]': '"var(--mat-sys-" + color() + ")"',
   },
 })
 export class SbDarkModeToggleMatComponent {
@@ -43,26 +42,23 @@ export class SbDarkModeToggleMatComponent {
    * Emits an event when the dark mode is toggled.
    * The event payload is a boolean indicating whether dark mode is enabled or not.
    */
-  _isDarkEvent = output<boolean>({ alias: 'toggleIsDark' })
 
-  /**
- * Controls whether to hide the switch UI element.
- * When true, only the icon will be visible wit`h cursor:pointer set.
- * @default true
- * @internal Use the [hideSwitch] template binding - do not access this property directly
- */
-  _hideSwitch = input(true, { alias: 'hideSwitch' })
-  _color = input<SbThemePalette>('primary', { alias: 'color' })
+  // Output: toggleIsDark
+  toggleIsDark = output<boolean>();
+
+  // Inputs: hideSwitch, color
+  hideSwitch = input(true);
+  color = input<SbThemePalette | undefined>(undefined);
 
   //- - - - - - - - - - - - - - -//
 
-  protected _isDark = this._themeService.isDarkMode
+  protected isDark = this._themeService.isDarkMode
 
   //- - - - - - - - - - - - - - -//
 
   toggleDarkTheme = (isDark: boolean) => {
-    this._themeService.setDarkMode(isDark ? 'dark' : 'light')
-    this._isDarkEvent.emit(isDark)
+    this._themeService.setDarkMode(isDark ? 'dark' : 'light');
+    this.toggleIsDark.emit(isDark);
   }
 
 } //Cls