+++ libs/utils/forms/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [2.0.1] - 2025-08-05
+
+### Changed
+- Added FirstErrorMatComponent
+
 ## [2.0.0] - 2025-08-01
 
 ### Changed